+++ sdks/python/src/opik/integrations/adk/llm_response_wrapper.py
@@ -1,6 +1,5 @@
 import dataclasses
 import logging
-import os
 from typing import Callable, Optional, Dict, Any
 
 from google.adk import models as adk_models
@@ -8,7 +7,6 @@ from google.genai import types as genai_types
 
 from ... import llm_usage, LLMProvider
 from ...llm_usage import opik_usage
-
 
 LOGGER = logging.Logger(__name__)
 
@@ -80,11 +78,3 @@ def _wrap_llm_response_create(
     response.custom_metadata["model_version"] = generate_content_response.model_version
 
     return response
-
-
-def get_adk_provider() -> LLMProvider:
-    use_vertexai = os.environ.get("GOOGLE_GENAI_USE_VERTEXAI", "0").lower() in [
-        "true",
-        "1",
-    ]
-    return LLMProvider.GOOGLE_VERTEXAI if use_vertexai else LLMProvider.GOOGLE_AI
